+++ packages/orq-rc/src/models/operations/createannotationqueue.ts
@@ -151,7 +151,7 @@ export const CreateAnnotationQueueResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-08-24T01:56:57.966Z",
+    "2026-08-24T09:28:39.484Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {