+++ packages/orq-rc/src/models/operations/updatetool.ts
@@ -1764,7 +1764,7 @@ export const UpdateToolResponseBodyCodeExecutionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M0RQRP36XQH1TE1NQXNJYXVC"),
+  _id: z.string().default("tool_01M0SHKRNQYBEG5RB6JM9CRTEX"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -1959,7 +1959,7 @@ export const UpdateToolResponseBodyHTTPTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M0RQRP3411GS16N8PFPWG36F"),
+  _id: z.string().default("tool_01M0SHKRNP5DE71D4GJSFZ3H75"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2054,7 +2054,7 @@ export const UpdateToolResponseBodyJSONSchemaTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M0RQRP343TBHA89CHYE6WC00"),
+  _id: z.string().default("tool_01M0SHKRNHXXTC3PVV1KTFXJJV"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2157,7 +2157,7 @@ export const UpdateToolResponseBodyFunctionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M0RQRP33QTAXEXMREB80S61N"),
+  _id: z.string().default("tool_01M0SHKRNG63SS4TX9N94N7T8G"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),